Fix some warning & refactory namespace
