+++ src/BeeNet/DtoModel/PendingTransactionDto.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Etherna.BeeNet.DtoModel.DebugApi
+namespace Etherna.BeeNet.DtoModel
 {
     public class PendingTransactionDto
     {
