+++ src/BeeNet/DtoModel/SettlementDataDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Etherna.BeeNet.DtoModel.DebugApi
+﻿namespace Etherna.BeeNet.DtoModel
 {
     public class SettlementDataDto
     {
