+++ src/BeeNet/DtoModel/SettlementDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Etherna.BeeNet.DtoModel.DebugApi
+namespace Etherna.BeeNet.DtoModel
 {
     public class SettlementDto
     {
