+++ src/BeeNet/DtoModel/TagsDto.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Etherna.BeeNet.DtoModel.DebugApi
+namespace Etherna.BeeNet.DtoModel
 {
     public class TagsDto
     {
